switch to fp16 as the go-to float dtype (i.e. how to infer a python float literal, or how to convert ints to floats). to try and simplify the conversion process (which currently starts by reading an fp16 torchscript scriptmodule, converts to fp32 CoreML, then optimizes back to fp16 CoreML). conversion completes, but XCode complains that the .mlpackage is missing inputSchema, and when you try to load the model in python you get complaints due to the changes made to type domains on upsample function

+++ coremltools/converters/mil/mil/ops/defs/iOS15/control_flow.py
@@ -160,7 +160,7 @@ class Const(Operation):
     def _get_type_val(self, value):
 
         if isinstance(value, (float, np.float64)):
-            value = np.float32(value)
+            value = np.float16(value)
         elif isinstance(value, bool):
             value = np.bool(value)
         elif isinstance(value, (int, np.int64)):
@@ -180,7 +180,7 @@ class Const(Operation):
             elif value.dtype == np.float64:
                 msg = "Downcast const op {} data fp64 as fp32".format(self.name)
                 logger.debug(msg)
-                value = value.astype(np.float32)
+                value = value.astype(np.float16)
 
         elif isinstance(value, mil_list):
             # if val that was passed in is of type mil_list, which is just a wrapper on top of python list
